Refactoring Implementation
