+++ src/main/java/ifpb/collections/Header.java
@@ -1,18 +1,19 @@
-package ifpb.entitycomplex;
+package ifpb.collections;
 
 import ifpb.collections.interfaces.IAuthors;
 import ifpb.entitybasic.exceptions.InvalidNullException;
 import ifpb.entitybasic.interfaces.IInformation;
 import ifpb.entitycomplex.exceptions.InvalidSizeAuthors;
-import ifpb.entitycomplex.interfaces.IHeader;
+import ifpb.collections.interfaces.IHeader;
 
 public class Header implements IHeader {
     private IInformation information;
     private IAuthors authors;
 
     public Header(IInformation information, IAuthors authors) throws Throwable {
-        this.information = validateInformation(information);
-        this.authors = validateSizeAuthors(authors);
+        this.information = hasNullInformation(information);
+        this.authors = hasNullAuthors(authors);
+        hasSizeValidForAuthor(authors);
     }
 
     @Override
@@ -25,7 +26,7 @@ public class Header implements IHeader {
         return this.authors;
     }
 
-    private IInformation validateInformation(IInformation information) throws InvalidNullException {
+    private IInformation hasNullInformation(IInformation information) throws InvalidNullException {
         if (information == null || information.getDescription() == null || information.getTitle() == null){
             throw new InvalidNullException();
         }
@@ -33,9 +34,12 @@ public class Header implements IHeader {
         return information;
     }
 
-    private IAuthors validateSizeAuthors(IAuthors authors) throws Throwable {
-        if (authors == null) throw new InvalidNullException();
+    private void hasSizeValidForAuthor(IAuthors authors) throws InvalidSizeAuthors {
         if (authors.get().length < 1) throw new InvalidSizeAuthors();
+    }
+
+    private IAuthors hasNullAuthors(IAuthors authors) throws Throwable {
+        if (authors == null) throw new InvalidNullException();
 
         return authors;
     }
